Hotfix: allow review labels with no score
While rare, it seems it is possible for a review label JSON entry
to contain no score. Whereas previously this would result in a
ValueError when we tried to access the json-parsed dictionary,
we now ignore label entries that do not contain a `value` key
when computing the resolved score and queue time.

super-commit: 27ec0ecdfd3a8bb426bdf11aa1ad1f6b9ec0a4da